[Feat] :Day4 강의 put,delete, 예외처리

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -61,7 +61,7 @@ public class UserController {
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request) {
         String readSql = "Select * from user where id = ?";
-        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, request.getId());
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, request.getId()).isEmpty();
         if (isUserNotExist) {
             throw new IllegalArgumentException();
         }
@@ -72,6 +72,11 @@ public class UserController {
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
+        String readSql = "Select * from user where name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
         String sql = "Delete from user where name = ?";
         jdbcTemplate.update(sql, name);
     }
